Remove std.prune from Job func
This prunes a bit too much; it gets rid of nulls, which is fine, but it
also gets rid of object members that have empty objects, e.g.:

```

{
  foo: {}
  bar: [0],
}
```

In the above, `foo` will be pruned. This is unfortunate because that
means services with Consul Connect enabled that have only an empty
SidecarService member will be pruned, and that is not desirable.

std.prune needs to be restored at a much lower level, if ever it is
restored.